[Fix] Voucher Code Warning (#232)
* Add checkoutUserErrors in Query

* Refactor ShoppingCart

* Add Components

* Voucher Code Warning
