+++ src/scenes/ShoppingCart/components/index.ts
@@ -1,0 +1,4 @@
+import ShoppingCartPayment from './ShoppingCartPayment';
+import BottomButton from './BottomButton';
+
+export { ShoppingCartPayment, BottomButton };
